revert: restore login as standalone page
Reverts the login modal approach back to the original dedicated login
page with AuthGuard redirect. Removes LoginModal.tsx.
