+++ src/components/AuthGuard.tsx
@@ -1,8 +1,16 @@
+import { useEffect } from 'react'
+import { useNavigate } from 'react-router-dom'
 import { useAuthStore } from '@/stores/authStore'
-import LoginModal from '@/components/LoginModal'
 
 export default function AuthGuard({ children }: { children: React.ReactNode }) {
   const { session, isLoading } = useAuthStore()
+  const navigate = useNavigate()
+
+  useEffect(() => {
+    if (!isLoading && !session) {
+      navigate('/login', { replace: true })
+    }
+  }, [isLoading, session, navigate])
 
   if (isLoading) {
     return (
@@ -13,7 +21,7 @@ export default function AuthGuard({ children }: { children: React.ReactNode }) {
   }
 
   if (!session) {
-    return <LoginModal open />
+    return null
   }
 
   return <>{children}</>
